refactor: ItemListRenderer -> ItemListResults로 컴포넌트명 변경 + 'apis' 폴더명 'services'로 변경
