+++ src/components/ItemList/BestItemList.jsx
@@ -2,7 +2,7 @@ import { useEffect, useState, useCallback } from "react";
 import { getItems } from "../../apis/api";
 import styles from "./ItemList.module.css";
 import useAsync from "../../hooks/useAsync";
-import ItemListRenderer from "./ItemListRenderer";
+import ItemListResults from "./ItemListResults";
 
 const ORDER_BY = "favorite";
 const LIST_TYPE = "best";
@@ -33,7 +33,7 @@ const BestItemList = ({ pageSize, title }) => {
   return (
     <div className={styles["item-list-area"]}>
       <h4 className={styles["item-list-title"]}>{title}</h4>
-      <ItemListRenderer
+      <ItemListResults
         isLoading={isLoading}
         isError={loadingError}
         items={items}
