+++ src/services/api.js
@@ -4,8 +4,9 @@ export async function getItems({
   page = 1,
   pageSize = 10,
   orderBy = "recent",
+  keyword = "",
 }) {
-  const query = `page=${page}&pageSize=${pageSize}&orderBy=${orderBy}`;
+  const query = `page=${page}&pageSize=${pageSize}&orderBy=${orderBy}&keyword=${keyword}`;
   const response = await fetch(`${BASE_URL}/products?${query}`);
 
   if (!response.ok) {
